+++ src/store/closure/mod.rs
@@ -423,6 +423,7 @@ mod tests {
             store_path: None,
             targets: Vec::new(),
             fixed_output: false,
+            build_only: false,
             summary: None,
             bins: BTreeMap::new(),
             sources: BTreeMap::new(),